requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
liac-arff==2.5.0
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
